Explicitly set the preferred route table
Since we need to use two subnets for redunancy (requirement to use
security groups) there is a chance the VPC will select the route table
associated to the secondary subnet, which isn't associated to the
gateway we need to access the instance.

Let's explicitly set the route table for the subnet that's associated
with the gateway to prevent random issues based on which subnet route
table AWS selects.