Add a basic database migration wrapper around golang-migrate/migrate
This commits introduces basic plumbing so we can start writing database
migrations. The script only supports migrating up, or applying
migrations. It doesn't support reverting a migration, or dropping
database. This can be done using the golang-migrate/migrate CLI directly
and a link to that documentation is included in the
migrations/README.md.